+++ public/tabs/chat_window_header.tsx
@@ -3,24 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {
-  EuiButtonEmpty,
-  EuiButtonIcon,
-  EuiContextMenuItem,
-  EuiContextMenuPanel,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiPopover,
-  EuiIcon,
-} from '@elastic/eui';
-import React, { useCallback, useState } from 'react';
-import { EditConversationNameModal } from '../components/edit_conversation_name_modal';
+import { EuiButtonIcon, EuiFlexGroup, EuiFlexItem, EuiIcon } from '@elastic/eui';
+import React from 'react';
 import { useChatContext } from '../contexts/chat_context';
-import { useChatActions } from '../hooks/use_chat_actions';
-import { NotebookNameModal } from '../components/notebook/notebook_name_modal';
-import { useCore } from '../contexts/core_context';
-import { useChatState } from '../hooks/use_chat_state';
-import { useSaveChat } from '../hooks/use_save_chat';
+import { ChatWindowHeaderTitle } from '../components/chat_window_header_title';
 import chatIcon from '../assets/chat.svg';
 interface ChatWindowHeaderProps {
   flyoutFullScreen: boolean;
@@ -29,30 +15,6 @@ interface ChatWindowHeaderProps {
 
 export const ChatWindowHeader: React.FC<ChatWindowHeaderProps> = React.memo((props) => {
   const chatContext = useChatContext();
-  const { loadChat } = useChatActions();
-  const core = useCore();
-  const [isPopoverOpen, setPopover] = useState(false);
-  const [isRenameModelOpen, setRenameModelOpen] = useState(false);
-  const { chatState } = useChatState();
-  const { saveChat } = useSaveChat();
-
-  const onButtonClick = () => {
-    setPopover(!isPopoverOpen);
-  };
-
-  const closePopover = () => {
-    setPopover(false);
-  };
-
-  const handleEditConversationClose = useCallback(
-    (status: 'updated' | string, newTitle?: string) => {
-      if (status === 'updated') {
-        chatContext.setTitle(newTitle);
-      }
-      setRenameModelOpen(false);
-    },
-    [chatContext]
-  );
 
   const dockBottom = () => (
     <svg width="16" height="16" viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg">
@@ -72,129 +34,61 @@ export const ChatWindowHeader: React.FC<ChatWindowHeaderProps> = React.memo((pro
     </svg>
   );
 
-  const button = (
-    <EuiButtonEmpty
-      style={{ maxWidth: '300px' }}
-      size="s"
-      color="text"
-      iconType="arrowDown"
-      iconSide="right"
-      onClick={onButtonClick}
-    >
-      <h3 className="eui-textTruncate">
-        {chatContext.sessionId ? chatContext.title : 'OpenSearch Assistant'}
-      </h3>
-    </EuiButtonEmpty>
-  );
-
-  const items = [
-    <EuiContextMenuItem
-      disabled={!chatContext.sessionId}
-      key="rename-conversation"
-      onClick={() => {
-        closePopover();
-        setRenameModelOpen(true);
-      }}
-    >
-      Rename conversation
-    </EuiContextMenuItem>,
-    <EuiContextMenuItem
-      key="new-conversation"
-      onClick={() => {
-        closePopover();
-        loadChat(undefined);
-      }}
-    >
-      New conversation
-    </EuiContextMenuItem>,
-    <EuiContextMenuItem
-      key="save-as-notebook"
-      onClick={() => {
-        const modal = core.overlays.openModal(
-          <NotebookNameModal onClose={() => modal.close()} saveChat={saveChat} />
-        );
-        closePopover();
-      }}
-      // User only can save conversation when he send a message at least.
-      disabled={chatState.messages.filter((item) => item.type === 'input').length < 1}
-    >
-      Save to notebook
-    </EuiContextMenuItem>,
-  ];
-
   return (
-    <>
-      <EuiFlexGroup
-        gutterSize="s"
-        justifyContent="spaceAround"
-        alignItems="center"
-        responsive={false}
-      >
-        <EuiFlexItem>
-          <EuiFlexGroup gutterSize="none" alignItems="center" responsive={false}>
-            <EuiFlexItem grow={false} style={{ marginLeft: '8px' }}>
-              <EuiIcon type={chatIcon} size="m" />
-            </EuiFlexItem>
-            <EuiFlexItem grow={false}>
-              <EuiPopover
-                id="conversationTitle"
-                button={button}
-                isOpen={isPopoverOpen}
-                closePopover={closePopover}
-                panelPaddingSize="none"
-                anchorPosition="downRight"
-              >
-                <EuiContextMenuPanel size="m" items={items} />
-              </EuiPopover>
-            </EuiFlexItem>
-            <EuiFlexItem grow={false}>
-              <EuiButtonIcon
-                aria-label="history"
-                iconType="clock"
-                size="xs"
-                color="text"
-                onClick={() => {
-                  chatContext.setFlyoutComponent(undefined);
-                  // Back to chat tab if history page already visible
-                  chatContext.setSelectedTabId(
-                    chatContext.selectedTabId === 'history' ? 'chat' : 'history'
-                  );
-                }}
-                display={chatContext.selectedTabId === 'history' ? 'fill' : undefined}
-              />
-            </EuiFlexItem>
-          </EuiFlexGroup>
-        </EuiFlexItem>
-        <EuiFlexItem grow={false}>
-          <EuiButtonIcon
-            aria-label="fullScreen"
-            color="text"
-            size="xs"
-            // TODO replace svg with built-in icon
-            iconType={props.flyoutFullScreen ? dockRight : dockBottom}
-            onClick={props.toggleFlyoutFullScreen}
-          />
-        </EuiFlexItem>
-        <EuiFlexItem grow={false}>
-          <EuiButtonIcon
-            aria-label="close"
-            size="xs"
-            color="text"
-            iconType="cross"
-            onClick={() => {
-              chatContext.setFlyoutVisible(false);
-            }}
-          />
-        </EuiFlexItem>
-        <EuiFlexItem grow={false} />
-      </EuiFlexGroup>
-      {isRenameModelOpen && (
-        <EditConversationNameModal
-          sessionId={chatContext.sessionId!}
-          onClose={handleEditConversationClose}
-          defaultTitle={chatContext.title!}
+    <EuiFlexGroup
+      gutterSize="s"
+      justifyContent="spaceAround"
+      alignItems="center"
+      responsive={false}
+    >
+      <EuiFlexItem>
+        <EuiFlexGroup gutterSize="none" alignItems="center" responsive={false}>
+          <EuiFlexItem grow={false} style={{ marginLeft: '8px' }}>
+            <EuiIcon type={chatIcon} size="m" />
+          </EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <ChatWindowHeaderTitle />
+          </EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <EuiButtonIcon
+              aria-label="history"
+              iconType="clock"
+              size="xs"
+              color="text"
+              onClick={() => {
+                chatContext.setFlyoutComponent(undefined);
+                // Back to chat tab if history page already visible
+                chatContext.setSelectedTabId(
+                  chatContext.selectedTabId === 'history' ? 'chat' : 'history'
+                );
+              }}
+              display={chatContext.selectedTabId === 'history' ? 'fill' : undefined}
+            />
+          </EuiFlexItem>
+        </EuiFlexGroup>
+      </EuiFlexItem>
+      <EuiFlexItem grow={false}>
+        <EuiButtonIcon
+          aria-label="fullScreen"
+          color="text"
+          size="xs"
+          // TODO replace svg with built-in icon
+          iconType={props.flyoutFullScreen ? dockRight : dockBottom}
+          onClick={props.toggleFlyoutFullScreen}
         />
-      )}
-    </>
+      </EuiFlexItem>
+      <EuiFlexItem grow={false}>
+        <EuiButtonIcon
+          aria-label="close"
+          size="xs"
+          color="text"
+          iconType="cross"
+          onClick={() => {
+            chatContext.setFlyoutVisible(false);
+          }}
+        />
+      </EuiFlexItem>
+      <EuiFlexItem grow={false} />
+    </EuiFlexGroup>
   );
 });
